Split test builtins
